backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
mysqlconnector
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
requests
loguru
pytest
pytest-asyncio
black
ruff
fastapi-mail==1.4.1
aiosmtplib==2.0.2
google-generativeai
PyPDF2
